Separate API operation from controller
